% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_date_if_all.R
\name{as_date_if_all}
\alias{as_date_if_all}
\title{Conditionally convert to Date if all non-NA values match the given format}
\usage{
as_date_if_all(x, format = "\%m/\%d/\%Y")
}
\arguments{
\item{x}{A vector}

\item{format}{Date format string (default "\%m/\%d/\%Y")}
}
\value{
Date vector or original vector
}
\description{
Conditionally convert to Date if all non-NA values match the given format
}
