% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_datetime_lenient.R
\name{as_datetime_lenient}
\alias{as_datetime_lenient}
\title{Lenient datetime parsing using lubridate::parse_date_time}
\usage{
as_datetime_lenient(x, orders = c("mdy", "ymd", "dmy", "Ymd", "mdY"))
}
\arguments{
\item{x}{A vector}

\item{orders}{Character vector of date orders}
}
\value{
POSIXct vector or original vector
}
\description{
Lenient datetime parsing using lubridate::parse_date_time
}
