% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_logical_if_all.R
\name{as_logical_if_all}
\alias{as_logical_if_all}
\title{Conditionally convert to logical (supports TRUE/FALSE/T/F/Yes/No/Y/N)}
\usage{
as_logical_if_all(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Logical vector or original vector
}
\description{
Conditionally convert to logical (supports TRUE/FALSE/T/F/Yes/No/Y/N)
}
