% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_numeric_if_all.R
\name{as_numeric_if_all}
\alias{as_numeric_if_all}
\title{Conditionally convert to numeric if all non-NA values appear numeric}
\usage{
as_numeric_if_all(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Numeric vector or original vector unchanged
}
\description{
Conditionally convert to numeric if all non-NA values appear numeric
}
