% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_df.R
\docType{data}
\name{toy_df}
\alias{toy_df}
\title{Toy dataset for demonstrating typecleanR functions}
\format{
A tibble with 5 rows and 5 columns
}
\usage{
toy_df
}
\description{
Toy dataset for demonstrating typecleanR functions
}
\keyword{datasets}
